[diffusion] resolve LoRA weight sources deterministically
