--- python/sglang/multimodal_gen/checkpoint_inspection/__init__.py
@@ -1,1 +1,0 @@
-"""Checkpoint source discovery and metadata inspection."""
