+++ python/sglang/multimodal_gen/runtime/utils/hf_diffusers_utils.py
@@ -52,6 +52,7 @@ from sglang.multimodal_gen.runtime.utils.model_overlay import (
 from sglang.multimodal_gen.runtime.utils.quantization_utils import (
     normalize_flat_modelopt_quant_config,
 )
+from sglang.multimodal_gen.runtime.weight_source import resolve_weight
 from sglang.srt.environ import envs
 from sglang.srt.utils.hf_transformers import check_gguf_file
 from sglang.utils import is_in_ci
@@ -616,42 +617,66 @@ def maybe_download_lora(
     Returns:
         Local path to the model
     """
-    # Repositories often publish several adapter revisions side by side.  If a
-    # filename is pinned, do not download every weight before selecting it.
-    # Keep JSON metadata so PEFT's lora_alpha remains available.
-    allow_patterns = (
-        ["*.json", weight_name, f"**/{weight_name}"]
-        if weight_name is not None
-        else ["*.json", "*.safetensors", "*.bin"]
-    )
+    if envs.SGLANG_USE_MODELSCOPE.get():
+        allow_patterns = (
+            ["*.json", weight_name, f"**/{weight_name}"]
+            if weight_name is not None
+            else ["*.json", "*.safetensors", "*.bin"]
+        )
+        local_path = maybe_download_model(
+            model_name_or_path,
+            local_dir,
+            download,
+            is_lora=True,
+            allow_patterns=allow_patterns,
+        )
+        if os.path.isfile(local_path):
+            return local_path
+        if weight_name is not None:
+            target = os.path.join(local_path, weight_name)
+            if not os.path.isfile(target):
+                raise FileNotFoundError(
+                    f"Specified lora_weight_name '{weight_name}' not found in "
+                    f"{local_path}"
+                )
+            return target
+        guessed = _best_guess_weight_name(local_path, file_extension=".safetensors")
+        if guessed is None and current_platform.is_rocm():
+            guessed = _best_guess_weight_name(
+                model_name_or_path, file_extension=".safetensors"
+            )
+        return os.path.join(local_path, guessed)
 
+    resolved_weight = resolve_weight(model_name_or_path, weight_name=weight_name)
+    selected_file = resolved_weight.selected_file
+    if not selected_file.endswith(".safetensors"):
+        raise ValueError(
+            "Native diffusion LoRA loading requires a safetensors file, got "
+            f"{selected_file!r}"
+        )
+
+    source = resolved_weight.inventory.source
+    if source.kind == "local":
+        assert source.local_path is not None
+        if os.path.isfile(source.local_path):
+            return source.local_path
+        return os.path.join(source.local_path, selected_file)
+
+    assert source.repo_id is not None
     local_path = maybe_download_model(
-        model_name_or_path,
+        source.repo_id,
         local_dir,
         download,
         is_lora=True,
-        allow_patterns=allow_patterns,
+        allow_patterns=["*.json", selected_file],
+        revision=resolved_weight.inventory.resolved_revision or source.revision,
     )
-    # return directly if local_path is a file
-    if os.path.isfile(local_path):
-        return local_path
-
-    if weight_name is not None:
-        target = os.path.join(local_path, weight_name)
-        if not os.path.isfile(target):
-            raise FileNotFoundError(
-                f"Specified lora_weight_name '{weight_name}' not found in {local_path}"
-            )
-        return target
-
-    guessed = _best_guess_weight_name(local_path, file_extension=".safetensors")
-    # AMD workaround: PR 15813 changed from model_name_or_path to local_path,
-    # which can return None. Fall back to original behavior on ROCm.
-    if guessed is None and current_platform.is_rocm():
-        guessed = _best_guess_weight_name(
-            model_name_or_path, file_extension=".safetensors"
+    target = os.path.join(local_path, selected_file)
+    if not os.path.isfile(target):
+        raise FileNotFoundError(
+            f"Resolved LoRA weight {selected_file!r} was not downloaded to {local_path}"
         )
-    return os.path.join(local_path, guessed)
+    return target
 
 
 def verify_model_config_and_directory(model_path: str) -> dict[str, Any]:
